ServerTests: fix unit test failure on Linux
It turns out the 10th byte of a gzip file is different depending
on operating system.  In fact, the value of the 10th byte says
which operating system the inflate algorithm ran on.

So ignore this byte since we really don't care what operating system
we run the code on.