Client: add support for connections upgraded by servers
The user may optionally provide a function for the client to call
if the server decides to upgrade the connection protocol.  If a function
is provided, it's called immediately when the client receives the
response, during the client's data receiving delegate call, allowing
the user to replace the data receive and connection broken delegates
on the connection object without any chance of losing data.

The upgrade connection protocol delegate gets the HTTP response,
a handle to the connection, and any excess data already received by
the client that came immediately after the HTTP response.

The client releases all handles to connections that are upgraded,
before returning from its data receive delegate.

If no upgrade connection protocol delegate is given, and the server
decides to upgrade the connection protocol, the client will still
release the connection, preventing it from accidentally being
reused to send any subsequent request.