Test: Don't write to unused local variable when destruction
The bools connectionDestroyed aren't used and thus can be removed. Also,
the connection->onDestruction lambda would get executed when connection
gets out of the scope, which is at the end of the function. This causes
the AddressSanitizer (or any of the other sanitizers, I forgot) to
complain about invalid memory access.